\name{chromoR-package}
\alias{chromoR-package}
\alias{chromoR}
\docType{package}
\title{
Analysis of chromosomal interactions data (Hi-C data)
}
\description{
ChromoR combines wavelet change point with Bayes Factor, for useful correction, 
segmentation and comparison of Hi-C contact maps. 
It provides a user friendly software solution, addressing the entire statistical 
pipeline required for the analysis of chromosomal interactions data.
}
\details{
\tabular{ll}{
Package: \tab chromoR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-07\cr
License: \tab GPL-2\cr
}
For an easy start with chromoR, check out the documentation and examples for 
correctCIM, compareCIM and segmentCIM

See also http://www.cl.cam.ac.uk/~ys388/chromoR/ for more examples and data sets.
}
\author{
Yoli Shavit <ys388@cam.ac.uk>
}
\references{
http://www.cl.cam.ac.uk/~ys388/chromoR/
}
\keyword{ package }
