\name{imr90.1.obs}
\alias{imr90.1.obs}
\docType{data}
\title{
A non-corrected CIM of replicate 1 of IMR90, for chromosomes 1 and 2
}
\description{
A non corrected chromosomal interaction map (CIM) for replicate 1 of IMR90, 
for chromosomes 1 and 2, generated from a publicly available Hi-C data set [1].
This matrix was build with BuildCIM given Hi-C mapped positions, 
publicly available 

from http://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE35156
}
\usage{data(imr90.1.obs)}
\format{
   A matrix
  }
\references{
[1] J. R. Dixon, S. Selvaraj, F. Yue, A. Kim, Y. Li, Y. Shen, M. Hu, J. S. Liu 
and B. Ren, Nature, 2012, 485, 376.

See also http://www.cl.cam.ac.uk/~ys388/chromoR/ for the genome wide CIM.  
}
\examples{
data(imr90.1.obs)
}
\keyword{datasets}
