\name{seg.imr90.obs}
\alias{seg.imr90.obs}
\docType{data}
\title{
A segmentation for imr90.1.obs CIM
}
\description{
A data table definding the genomic coordinates of each row/col in imr90.1.obs 
(genome segmentation).
}
\usage{data(seg.imr90.obs)}
\format{
A data frame with 3 variables: chr, start, end
}
\examples{
data(seg.imr90.obs)
}
\keyword{datasets}
