\name{seg}
\alias{seg}
\docType{data}
\title{
A segmentation for imr90.1 and imr90.2 CIMs
}
\description{
A data table definding the genomic coordinates of each row/col in the corresponding CIMs (genome segmentation).
}
\usage{data(seg)}
\format{
  A data frame 3 variables: chr, start, end
}
\examples{
data(seg)
}
\keyword{datasets}
